Add callback support for un-referencing xprt user-data before destroy
When destroying an svc-xprt object, the references (to and from it)
held through its user-data fields must be released. Such unreleased
references to the svc-xprt object would otherwise not allow it to be
completely destroyed, resulting in leakage. Additionally, the
unreleased references to other data held by this svc-xprt object
would prevent that data to correctly use the svc-xprt due to it being
under-destruction.

This CL adds support for registering a callback function that would
perform the above un-referencing. This function would then be called
during destruction of an svc-xprt object.

Additionally, this CL also updates the existing `svc_destroy_it`
function (that is, the `SVC_DESTROY` macro) to call the above added
callback function.